为SVN数据上传客户端添加各种新的方法
